Change requirement from pycryptodome to pycryptodomex
rational: pycryptodome and pycrypto are imported as "Crypto". Installing both at once is a problem.
pycryptodomex is equivalent to pycryptodome, exceot it is imported as "Cryptodome".
The requirement was switched to pycryptodomex to break computers with pycrypto installed.

Despite the requirement list containing only one of the packages, damgard_jurik can use either of the three.